requirements.txt
absl-py==1.4.0
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.95.1
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
jellyfish==0.9.0
joblib==1.2.0
jupyter-client==7.3.4
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lib50==3.0.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
submit50==3.1.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.21.1
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
